+++ src/main/java/com/getpcpanel/cpp/linux/LinuxAudioSession.java
@@ -4,7 +4,6 @@ import java.io.File;
 
 import org.springframework.context.ApplicationEventPublisher;
 
-import com.getpcpanel.cpp.AudioDevice;
 import com.getpcpanel.cpp.AudioSession;
 
 import lombok.Getter;
@@ -13,8 +12,8 @@ import lombok.Getter;
 public class LinuxAudioSession extends AudioSession {
     private final int index;
 
-    public LinuxAudioSession(AudioDevice device, ApplicationEventPublisher eventPublisher, int index, int pid, File executable, String title, String icon, float volume, boolean muted) {
-        super(device, eventPublisher, pid, executable, title, icon, volume, muted);
+    public LinuxAudioSession(ApplicationEventPublisher eventPublisher, int index, int pid, File executable, String title, String icon, float volume, boolean muted) {
+        super(eventPublisher, pid, executable, title, icon, volume, muted);
         this.index = index;
     }
 }