+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Allow setting system sounds when the process does not have pid 0
 - It's possible to drag audio devices into the Selected Devices list in the Toggle Device button option again (fix for #2)
 - Updated JNativeHook so that dead keys keep working (fix for #4)
+- Applications that have multiple audio sessions should now be controlled correctly
 
 ## [1.3]